更改部分依赖
