+++ src/lib/static/css/main.css
@@ -34,5 +34,9 @@
 /*子页面*/
 .subclass {
     padding: 36px 10px 10px;
+    height: 100%;
+}
+
+#app .subclass{
     height: calc(100% - 50px);
 }
